reservation refund finished geonwook 23/01/09

+++ src/main/java/com/ictproject/wyhotel/reservation/service/ReservationServiceImpl.java
@@ -276,7 +276,25 @@ public class ReservationServiceImpl implements IReservationService {
 		
 		mapper.cancelDining(resvNum);
 		String memberCode = (String) session.getAttribute("member");
-		if (mapper.getReservRoomListOnlyCode(memberCode) == 0) {
+		if (mapper.getReservRoomListOnlyCode(memberCode) == 0
+			&& mapper.getReservDiningListOnlyCode(memberCode) == 0) {
+			mapper.deleteNmem(memberCode);
+		}
+	}
+	
+	@Override
+	public String getPaymentKey(String resvNum) {
+	
+		return mapper.getPaymentKey(resvNum);
+	}
+	
+	@Override
+	public void cancelRoom(String resvNum, HttpSession session) {
+	
+		String memberCode = (String) session.getAttribute("member");
+		mapper.cancelRoom(resvNum);
+		if (mapper.getReservRoomListOnlyCode(memberCode) == 0
+			&& mapper.getReservDiningListOnlyCode(memberCode) == 0) {
 			mapper.deleteNmem(memberCode);
 		}
 	}
